code/guiding-genai-bots-with-semantic-router/requirements.txt
semantic-router==0.0.32
python-dotenv==1.0.0
langchain==0.1.10
langchain-openai==0.0.8
langchainhub==0.1.15